adding gocmd flag to more easily test against go1.8beta2 and future builds
